+++ chats/apps/api/v1/accounts/viewsets.py
@@ -15,7 +15,6 @@ from chats.core.cache_utils import get_user_id_by_email_cached
     name="create", decorator=swagger_auto_schema(responses={201: '{"token":"TOKEN"}'})
 )
 class LoginViewset(mixins.CreateModelMixin, viewsets.GenericViewSet):
-
     """
     Login Users
     """
@@ -59,6 +58,6 @@ class UserDataViewset(viewsets.GenericViewSet):
         try:
             instance = self.get_object()
         except User.DoesNotExist:
-            return Response({"Detail": "Email not found"}, status.HTTP_404_NOT_FOUND)
+            return Response({"detail": "Email not found"}, status.HTTP_404_NOT_FOUND)
         serializer = self.get_serializer(instance)
         return Response(serializer.data)